chore: converge on a single int test
